requirements-data.txt
# pip install -r requirements-data.txt
google-cloud-storage>=2.10.0
python-docx>=1.1.0
